+++ packages/unmock-jest/src/reporter/index.ts
@@ -1,7 +1,7 @@
 // tslint:disable:no-console
 import { utils as unmockUtils } from "unmock";
 import { IReporterOptions, resolveOptions } from "./options";
-import writeReport, { IJestData } from "./report-writer";
+import writeReport, { IJestData } from "./write-report";
 
 /**
  * Write report from Jest test results, loading snapshots, building report and writing to file.
